added thin web server
Following http://devcenter.heroku.com/articles/rails3/#webserver, I added the thin web server as a dependency in the
gemfile and added the complementary Procfile at the application root.

The starting thin with $ foreman start does not yield the same output as shown at the heroku link. But with a little
cajoling, e.g., opening a new tab in chrome and finding the app at port 5000, it appears to work